requirements.txt
discord.py>=2.0.1
async-timeout>=4.0.1
asyncio>=3.4.3
PyNaCl>=1.4.0
DateTime>=4.7.0